+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/sharing/ShareFilesViewModel.kt
@@ -120,7 +120,8 @@ class SharedDataViewModelImpl(
                         val size = file.fileSize
                         val (width, height) = if (size == null || size <= maxMediaSize) get<GetImageDimensions>().invoke(
                             file.content,
-                            maxMediaSize
+                            maxMediaSize,
+                            file.mimeType
                         ) else Pair(null, null)
                         image(
                             body = file.fileName,